pyproject.toml
[project]
name = "fitnessllm_dataplatform"
version = "0.1.0"
description = ""
authors = [
    { name = "Santosh G", email = "santosh@santoshg.io" }
]

[tool.poetry.dependencies]
python = "3.12.2"
cloudpathlib = {extras = ["gs"], version = "^0.20.0"}
typing-extensions = "^4.12.2"
fire = "^0.7.0"
google-cloud-secret-manager = "^2.22.1"
tqdm = "^4.67.1"
numpy = "^2.2.2"
pandas = "^2.2.3"
google-cloud-bigquery = {extras = ["pandas"], version = "^3.29.0"}
joblib = "^1.4.2"
tqdm-joblib = "^0.0.4"
google-cloud-bigquery-storage = "^2.27.0"
stravalib = "^2.1"
redis = "^5.2.1"
pytest = "^8.3.4"
beartype = "^0.19.0"
pytest-cov = "^6.0.0"
pre-commit = "^4.1.0"
jinja2 = "^3.1.5"
sqlglot = "^26.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
